requirements.txt
flask
flask-cors
flask-mysqldb
mysql-connector-python-rf
python-dotenv
wheel
pytest
flask-bcrypt